requirements.txt
Flask==3.0.1
Flask-Cors==4.0.0
poke-env
python-dotenv
